Implementación del custom hook con fetch
