+++ src/hooks/useFetchGifs.js
@@ -1,0 +1,24 @@
+// Hook es una función que regresa algo
+
+import { useEffect, useState } from "react";
+import { getGifs } from "../helpers/getGifs";
+
+export const useFetchGifs = (category) => {
+  const [images, setImages] = useState([]);
+  const [isLoading, setIsLoading] = useState(true);
+
+  const getImages = async () => {
+    const newImages = await getGifs(category);
+    setImages(newImages);
+    setIsLoading(false);
+  };
+
+  useEffect(() => {
+    getImages();
+  }, []);
+
+  return {
+    images,
+    isLoading,
+  };
+};
